classroom/pyproject.toml
[tool.poetry]
name = "classroom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
google = "^3.0.0"
google-auth-oauthlib = "^1.1.0"
google-auth-stubs = "^0.2.0"
prisma = "^0.11.0"
cryptography = "^41.0.7"


[tool.poetry.group.dev.dependencies]
google-auth-oauthlib-stubs = {git = "https://github.com/cffnpwr/google-auth-oauthlib-stubs.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
